Balance component + tweaks to zone size
