+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: planet_city_builder
 description: "CS 545 Jahnu Best."
 publish_to: 'none'
-version: 0.1.1
+version: 0.1.2
 
 environment:
   sdk: ^3.5.3
@@ -13,6 +13,8 @@ dependencies:
   permission_handler: ^11.0.0
   file_picker: ^8.0.0
   flame: ^1.21.0
+  intl: ^0.18.0
+  package_info_plus: ^4.0.0
 
 dev_dependencies:
   flutter_test:
